requirements.txt

fastapi
uvicorn
pydantic
openai-whisper
deep-translator
